requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.10.6
pydantic-settings==2.8.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
python-dotenv==1.0.1
requests==2.31.0
openai==1.12.0
regex==2023.12.25
cryptography==42.0.2